Reduce reviewer sanitizer drift
